requirements.txt
certifi==2025.4.26
contourpy==1.3.2
cycler==0.12.1
ExifRead==3.1.0
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
packaging==25.0
piexif==1.1.3
pillow==11.2.1
pycocotools==2.0.8
pyparsing==3.2.3
pyproj==3.7.1
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.0
python-dateutil==2.9.0.post0
scipy==1.15.2
setuptools==79.0.1
six==1.17.0
sympy==1.13.3
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
triton==3.3.0
typing_extensions==4.13.2
